fuck go back (reverting the merge)
This reverts commit 274f3598f8c00194b6cd9d0ce09d8d5cf71df694, reversing
changes made to 50944c8f0d64beaf35054fec68fb3bc711cbbcc8.

+++ code/modules/food_and_drinks/restaurant/_venue.dm
@@ -223,14 +223,15 @@
 	linked_venue.toggle_open()
 	update_icon()
 
-/obj/machinery/restaurant_portal/item_interaction(mob/living/user, obj/item/tool, list/modifiers)
-	var/obj/item/card/id/used_id = tool.GetID()
-	if(!istype(used_id))
-		return NONE
+/obj/machinery/restaurant_portal/attacked_by(obj/item/I, mob/living/user)
+	if(!istype(I,  /obj/item/card/id))
+		return ..()
 
-	if(!check_access(used_id))
-		balloon_alert(user, "insufficient access!")
-		return ITEM_INTERACT_BLOCKING
+	var/obj/item/card/id/used_id = I
+
+	if(!(linked_venue.req_access in used_id.GetAccess()))
+		to_chat(user, span_warning("This card lacks the access to change this venues status."))
+		return
 
 	var/list/radial_items = list()
 	var/list/radial_results = list()
@@ -243,7 +244,7 @@
 	var/choice = show_radial_menu(user, src, radial_items, null, require_near = TRUE)
 
 	if(!choice)
-		return ITEM_INTERACT_BLOCKING
+		return
 
 	var/venue_type = radial_results[choice]
 	var/obj/item/circuitboard/machine/restaurant_portal/board = circuit
@@ -252,7 +253,11 @@
 
 	turned_on_portal = WEAKREF(user)
 
-	balloon_alert(user, "venue changed to [chosen_venue.name]")
+	if(!(chosen_venue.req_access in used_id.GetAccess()))
+		to_chat(user, span_warning("This card lacks the access to change this venues status."))
+		return
+
+	to_chat(user, span_notice("You change the portal's linked venue."))
 
 	if(linked_venue && (src in linked_venue.restaurant_portals)) //We're already linked, unlink us.
 		linked_venue.restaurant_portals -= src
@@ -262,7 +267,6 @@
 
 	linked_venue = chosen_venue
 	linked_venue.restaurant_portals += src
-	return ITEM_INTERACT_SUCCESS
 
 /obj/machinery/restaurant_portal/screwdriver_act(mob/user, obj/item/tool)
 	if (default_deconstruction_screwdriver(user, "[base_icon_state]-open", base_icon_state, tool))
